# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AUK Lost + Found API
  description: |
    REST API for the lost-and-found system (CPEG/CSIS445 Assignment 2, Theme 14).
    Deployed on AWS: web server on EC2, data in Amazon RDS (MySQL), item photos
    in a private Amazon S3 bucket accessed only through pre-signed URLs.
  version: 2.0.0
servers:
  - url: http://EC2_PUBLIC_IP:3000/api
    description: AWS deployment (EC2)
  - url: http://localhost:3000/api
    description: Local development
paths:
  /uploads:
    post:
      summary: Get a pre-signed S3 upload URL for one item photo
      description: |
        Returns a short-lived pre-signed PUT URL. The client uploads the image
        bytes directly to S3 with `PUT upload_url` (Content-Type must match),
        then passes the returned `key` as `photo_key` when creating or
        updating an item. The bucket itself is private; these URLs are the
        only way in or out.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type]
              properties:
                content_type:
                  type: string
                  enum: [image/jpeg, image/png, image/webp]
      responses:
        '201':
          description: Pre-signed upload URL issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string, example: "items/3f8e9c2a-6f0d-4b7e-9a1c-2d4e5f6a7b8c.jpg" }
                  upload_url: { type: string, description: "Pre-signed S3 PUT URL (valid 5 minutes)" }
                  expires_in: { type: integer, example: 300 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503':
          description: S3 is not configured on this server
  /items:
    get:
      summary: List items
      parameters:
        - { name: type, in: query, schema: { type: string, enum: [lost, found] } }
        - { name: category, in: query, schema: { $ref: '#/components/schemas/Category' } }
        - { name: status, in: query, schema: { type: string, enum: [open, resolved] } }
        - { name: q, in: query, description: "Text search over title, description and location", schema: { type: string } }
      responses:
        '200':
          description: Array of items, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Item' } }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      summary: Report a lost or found item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemInput' }
      responses:
        '201':
          description: Item created
          headers:
            Location: { schema: { type: string }, description: URL of the new item }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /items/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      summary: Get one item
      responses:
        '200':
          description: The item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update an item (partial updates allowed)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemUpdate' }
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete an item
      responses:
        '204': { description: "Deleted, no content in the response" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  schemas:
    Category:
      type: string
      enum: [electronics, clothing, documents, keys, bags, jewelry, books, other]
    Item:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        type: { type: string, enum: [lost, found] }
        title: { type: string }
        description: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        location: { type: string }
        date: { type: string, format: date }
        status: { type: string, enum: [open, resolved] }
        contact_name: { type: string }
        contact_info: { type: string }
        photo_key:
          type: string
          nullable: true
          description: S3 object key of the item photo (null if none)
        photo_url:
          type: string
          nullable: true
          readOnly: true
          description: Pre-signed S3 GET URL for the photo (valid 1 hour, null if no photo)
        created_at: { type: string, readOnly: true }
    ItemInput:
      type: object
      required: [type, title, category, location, date, contact_name, contact_info]
      properties:
        type: { type: string, enum: [lost, found] }
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        location: { type: string, minLength: 1 }
        date: { type: string, format: date }
        contact_name: { type: string, minLength: 1 }
        contact_info: { type: string, minLength: 1 }
        photo_key:
          type: string
          nullable: true
          description: Key returned by POST /uploads (optional)
    ItemUpdate:
      description: Any subset of ItemInput fields, plus status.
      type: object
      properties:
        type: { type: string, enum: [lost, found] }
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        location: { type: string, minLength: 1 }
        date: { type: string, format: date }
        status: { type: string, enum: [open, resolved] }
        contact_name: { type: string, minLength: 1 }
        contact_info: { type: string, minLength: 1 }
        photo_key:
          type: string
          nullable: true
          description: New photo key from POST /uploads, or null to remove the photo
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              errors: { type: array, items: { type: string } }
    NotFound:
      description: Item not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }